report/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-crispy-forms==2.3
et_xmlfile==2.0.0
Faker==35.0.0
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
